requirements.txt
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Bootstrap-Flask==2.2.0
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
dominate==2.9.1
email_validator==2.2.0
Flask==2.3.3
Flask-CKEditor==1.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
libgravatar==1.0.4
MarkupSafe==3.0.2
numpy==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
spotipy==2.24.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
visitor==0.1.3
webencodings==0.5.1
Werkzeug==2.3.8
WTForms==3.1.2
